diag-parser: add support for non-Redhat variants
If there is no /etc/redhat-release file we pick up the information
from the lsb_release commandline.

In addition, we also retrieve (if present) the value of DKUDISTRIB which
DSS sets at install time (but not always). It allows to handle cases where
neither of the abovementioned information is present (e.g. Amazon Linux or
SuSe) or incomplete (RHEL variants like Oracle show up as RHEL in
/etc/redhat-release).

The real solution would probably be to pick up the information from
/etc/os-release but it is not currently present in the diags.
A Trello has been filed to that end at:
https://trello.com/c/5gY0Z9ka/10291-vincent-rubiolo-add-etc-os-release-dump-into-dss-diags

Tested with: CentOS, Almalinux, RHEL, Debian, Ubuntu, SuSe